new commit

+++ src/main/java/com/app/request/service/RequestService.java
@@ -12,7 +12,6 @@ import com.app.request.result.Success;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Function;
@@ -30,7 +29,7 @@ public class RequestService {
         Result result = null;
         try {
             if (itemQuantity <= 0)
-                throw new InvalidItemQuantityException(translate("error_valid_address"));
+                throw new InvalidItemQuantityException(translate("error_valid_quantity"));
             if (mediaList != null && mediaList.length > 3)
                 throw new InvalidMediaSizeException(translate("error_valid_upload"));
             result = Optional.ofNullable(repositoryImpl.addRequestInBucket(itemId, itemQuantity, mediaList, customerId)
@@ -54,30 +53,35 @@ public class RequestService {
     }
 
     public Result deleteMediaById(String customerId, long requestId, List<String> mediaIds) {
-       return Optional.ofNullable(repositoryImpl.deleteMedia(customerId, requestId, mediaIds))
-               .map((Function<List<Media>, Result<List<Media>>>) mediaList -> new Success<>(mediaList))
-               .orElse(new ServerError(new Exception(translate("error_something_went_wrong"))));
+        Result result = null;
+        try {
+            int size = repositoryImpl.getAvailableMediaListSizeForRequestItem(requestId, customerId);
+            int totalSize = mediaIds.size() > size ? mediaIds.size() - size : size - mediaIds.size();
+            if (totalSize == 0)
+                throw new InvalidMediaSizeException(translate("error_valid_delete"));
+            result = Optional.ofNullable(repositoryImpl.deleteMedia(customerId, requestId, mediaIds))
+                    .map((Function<List<Media>, Result<List<Media>>>) mediaList -> new Success<>(mediaList))
+                    .orElse(new ServerError(new Exception(translate("error_something_went_wrong"))));
+        } catch (InvalidMediaSizeException e) {
+            result = new ClientError(e);
+        }
+        return result;
     }
 
     public Result updateRequestInBucket(long requestItemId, int itemQuantity, MultipartFile[] mediaList, String customerId) {
-        /*try {
+        Result result = null;
+        try {
             if (itemQuantity <= 0)
-                throw new InvalidItemQuantityException();
-            int availableMediaSize = repositoryImpl.getAvailableMediaListSizeForRequestItem(requestItemId);
-            if (mediaList != null && (mediaList.length + availableMediaSize) > 3)
-                throw new InvalidMediaSizeException();
-            return Optional.ofNullable(
-                    repositoryImpl.updateRequestInBucket(requestItemId, itemQuantity, mediaList, customerId)
-            ).map(requestItem -> new Success("Request is Updated Successfully.")
-            ).get();
-        } catch (InvalidItemQuantityException e) {
-            return new ClientError(e.getMessage());
-        } catch (InvalidMediaSizeException e) {
-            return new ClientError(e.getMessage());
-        } catch (NoRequestItemFoundException e) {
-            return new ClientError(e.getMessage());
-        }*/
-        return null;
+                throw new InvalidItemQuantityException(translate("error_valid_quantity"));
+            if (mediaList != null && (mediaList.length + repositoryImpl.getAvailableMediaListSizeForRequestItem(requestItemId, customerId)) > 3)
+                throw new InvalidMediaSizeException(translate("error_valid_upload"));
+            result = Optional.ofNullable(repositoryImpl.updateRequestInBucket(requestItemId, itemQuantity, mediaList, customerId))
+                    .map((Function<RequestItem, Result<RequestItem>>) requestItem -> new Success<>(requestItem))
+                    .orElse(new ServerError(new Exception(translate("error_something_went_wrong"))));
+        } catch (InvalidItemQuantityException | InvalidMediaSizeException e) {
+            result = new ClientError(e);
+        }
+        return result;
     }
 
     public Result deleteRequestsInBucket(String customerId, List<Long> requestIds) {
